requirements.txt
yfinance>=0.2.31
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
numpy>=1.24.0
mcp>=1.0.0
google-genai>=1.0.0
streamlit>=1.30.0
pandas>=2.0.0
